third_party/WebKit/LayoutTests/webexposed/permissions-attribute-expected.txt

CONSOLE ERROR: line 7: Error while parsing the 'permissions' attribute: 'abcdefg' is an invalid permissions flag.
Tests that the permissions attribute is not exposed for stable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
